+++ packages/eslint-plugin/src/rules/check-sql.worker.ts
@@ -90,6 +90,7 @@ function workerHandler(params: WorkerParams): TE.TaskEither<WorkerError, WorkerR
         query: params.query,
         cacheKey: databaseUrl,
         pgParsed: params.pgParsed,
+        overrides: params.connection.overrides,
       });
     }),
     TE.chainW(TE.fromEither)
