+++ packages/shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./common";
 export * from "./errors";
 export * from "./parsedQuery";
+export * from "./pg";
